Moved getSecurityOpts() function to OS specific manager variant
