+++ pkg/kubelet/dockertools/docker_manager_linux.go
@@ -1,6 +1,9 @@
 package dockertools
 
-import dockertypes "github.com/docker/engine-api/types"
+import (
+	dockertypes "github.com/docker/engine-api/types"
+	"k8s.io/kubernetes/pkg/api"
+)
 
 func getContainerIP(container *dockertypes.ContainerJSON) string {
 	result := ""
@@ -21,4 +24,22 @@ func getNetworkingMode() string { return "" }
 // Returns true if the container name matches the infrastructure's container name
 func containerProvidesPodIP(name *KubeletContainerName) bool {
 	return name.ContainerName == PodInfraContainerName
+}
+
+// Returns Seccomp and AppArmor Security options
+func (dm *DockerManager) getSecurityOpts(pod *api.Pod, ctrName string) ([]dockerOpt, error) {
+	var securityOpts []dockerOpt
+	if seccompOpts, err := dm.getSeccompOpts(pod, ctrName); err != nil {
+		return nil, err
+	} else {
+		securityOpts = append(securityOpts, seccompOpts...)
+	}
+
+	if appArmorOpts, err := dm.getAppArmorOpts(pod, ctrName); err != nil {
+		return nil, err
+	} else {
+		securityOpts = append(securityOpts, appArmorOpts...)
+	}
+
+	return securityOpts, nil
 }
