+++ pkg/kubelet/dockertools/docker_manager_windows.go
@@ -3,6 +3,8 @@ package dockertools
 import (
 	"os"
 
+	"k8s.io/kubernetes/pkg/api"
+
 	dockertypes "github.com/docker/engine-api/types"
 )
 
@@ -30,4 +32,9 @@ func getNetworkingMode() string {
 // make sure to not grab the infra container's IP for the pod.
 func containerProvidesPodIP(name *KubeletContainerName) bool {
 	return name.ContainerName != PodInfraContainerName
+}
+
+// Returns nil as both Seccomp and AppArmor security options are not valid on Windows
+func (dm *DockerManager) getSecurityOpts(pod *api.Pod, ctrName string) ([]dockerOpt, error) {
+	return nil, nil
 }
